[Keyboard] Dumbpad v3.x review findings
