--- keyboards/dumbpad/v3x/templates/keymap.c
@@ -1,27 +1,0 @@
-#include QMK_KEYBOARD_H
-
-const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {__KEYMAP_GOES_HERE__};
-
-bool encoder_update_user(uint8_t index, bool clockwise) {
-    if (index == 0) {
-        switch (get_highest_layer(layer_state)) {
-            case 0:
-                // main layer, volume
-                if (clockwise) {
-                    tap_code(KC_VOLU);
-                } else {
-                    tap_code(KC_VOLD);
-                }
-                break;
-            default:
-                // rgb control layer, effects
-                if (clockwise) {
-                    rgblight_step();
-                } else {
-                    rgblight_step_reverse();
-                }
-                break;
-        }
-    }
-    return true;
-}
