+++ keyboards/dumbpad/v3x/v3x.c
@@ -94,6 +94,30 @@ void matrix_init_kb(void) {
     matrix_init_user();
 }
 
+bool encoder_update_user(uint8_t index, bool clockwise) {
+    if (index == 0) {
+        switch (get_highest_layer(layer_state)) {
+            case 0:
+                // main layer, volume
+                if (clockwise) {
+                    tap_code(KC_VOLU);
+                } else {
+                    tap_code(KC_VOLD);
+                }
+                break;
+            default:
+                // rgb control layer, effects
+                if (clockwise) {
+                    rgblight_step();
+                } else {
+                    rgblight_step_reverse();
+                }
+                break;
+        }
+    }
+    return true;
+}
+
 void led_set_kb(uint8_t usb_led) {
     // put your keyboard LED indicator (ex: Caps Lock LED) toggling code here
     writePin(LED_02, !IS_LED_ON(usb_led, USB_LED_NUM_LOCK));
